Add cypress tests, refactor updating, etc....
Cypress tests

Cypress is newly installed, so the CLI now works, Also the MSW were enabled as server listeners. The tsconfig was also added to resolve the "cy is undefined".

Updating the my articles

The update it self is changed when original array of articles before sort is updated. Right now the deleteArticleThunk is not returning also the nowSort array, but useEffect in EditArticleRowButtons was added. It listens for the originalArray prop and whenever the prop is updated, the useEffect creates new sort array with isActive state.

Unit and integration tests

Few tests and components were updated, so the coverage is almost green, except the storyshots which needs to be solved in future.