Delete cypress tests, add typing for modules and overview the FIXME/TODO
Cypress tests
Cypress tests deleted because of the wrong usage not compliant to the best practises. Non-logical rerendering with the it blocks

Typing for module
The JSONResolve enabled in the tsconfig.json file to resolve json files. Also declare module added to the index.d.ts

Misc
Revisited the TODO tree, to remove the already solved things