Connect the grapqhl image data to the app
In this case, the blob that was originally downloaded from the API couldn't be processed to show it in the app with the help of the URL.createObjectURL() where the blob file type is needed.

The solution to this was the responseType set to arraybuffer, which was converted to the Base64 string in gatsby-node. In the app, I've just added the base64 prefix for the rendering